general : debugging

+++ src/main/scala/com/muhalfian/spark/util/AggTools.scala
@@ -34,7 +34,7 @@ object AggTools {
   val readConfig = ReadConfig(Map("uri" -> "mongodb://10.252.37.112/prayuga", "database" -> "prayuga", "collection" -> "master_word_2"))
   var masterWord = MongoSpark.load(spark, readConfig).select("word", "index").map(row => {
     // var data = row.getAs[String](0)
-    (row.getAs[String](0),row.getAs[Double](1))
+    (row.getAs[String](0),row.getAs[Integer](1))
   }).collect.toArray//.map(_.toSeq)
   .foreach(println)
   // var masterWordCount = masterWord.size
